+++ examples/Makefile.am
@@ -16,38 +16,43 @@ if BUILD_ANIMDIFF
   noinst_PROGRAMS = anim_diff
 endif
 
+noinst_LTLIBRARIES = libexample_util.la
+
+libexample_util_la_SOURCES = example_util.c example_util.h
+libexample_util_la_LIBADD = ../src/libwebp.la
+
 anim_diff_SOURCES = anim_diff.c anim_util.c anim_util.h
 anim_diff_CPPFLAGS = $(AM_CPPFLAGS) $(USE_EXPERIMENTAL_CODE) $(GIF_INCLUDES)
 anim_diff_LDADD  = ../src/demux/libwebpdemux.la
-anim_diff_LDADD += ../imageio/libexample_util.la ../imageio/libimageio_util.la
+anim_diff_LDADD += libexample_util.la ../imageio/libimageio_util.la
 anim_diff_LDADD += $(GIF_LIBS) -lm
 
 dwebp_SOURCES = dwebp.c stopwatch.h
 dwebp_CPPFLAGS  = $(AM_CPPFLAGS) $(USE_EXPERIMENTAL_CODE)
 dwebp_CPPFLAGS += $(JPEG_INCLUDES) $(PNG_INCLUDES)
-dwebp_LDADD  = ../imageio/libexample_util.la ../imageio/libimageio_util.la
+dwebp_LDADD  = libexample_util.la ../imageio/libimageio_util.la
 dwebp_LDADD += ../imageio/libimagedec.la ../src/libwebp.la
 dwebp_LDADD +=$(PNG_LIBS) $(JPEG_LIBS)
 
 cwebp_SOURCES  = cwebp.c stopwatch.h
 cwebp_CPPFLAGS  = $(AM_CPPFLAGS) $(USE_EXPERIMENTAL_CODE)
-cwebp_LDADD  = ../imageio/libexample_util.la ../imageio/libimageio_util.la
+cwebp_LDADD  = libexample_util.la ../imageio/libimageio_util.la
 cwebp_LDADD += ../imageio/libimagedec.la ../src/libwebp.la
 cwebp_LDADD += $(JPEG_LIBS) $(PNG_LIBS) $(TIFF_LIBS)
 
 gif2webp_SOURCES = gif2webp.c gifdec.c gifdec.h
 gif2webp_CPPFLAGS = $(AM_CPPFLAGS) $(USE_EXPERIMENTAL_CODE) $(GIF_INCLUDES)
-gif2webp_LDADD  = ../imageio/libexample_util.la ../imageio/libimageio_util.la
+gif2webp_LDADD  = libexample_util.la ../imageio/libimageio_util.la
 gif2webp_LDADD += ../src/mux/libwebpmux.la ../src/libwebp.la $(GIF_LIBS)
 
 webpmux_SOURCES = webpmux.c
 webpmux_CPPFLAGS = $(AM_CPPFLAGS) $(USE_EXPERIMENTAL_CODE)
-webpmux_LDADD  = ../imageio/libexample_util.la ../imageio/libimageio_util.la
+webpmux_LDADD  = libexample_util.la ../imageio/libimageio_util.la
 webpmux_LDADD += ../src/mux/libwebpmux.la ../src/libwebp.la
 
 vwebp_SOURCES = vwebp.c
 vwebp_CPPFLAGS = $(AM_CPPFLAGS) $(USE_EXPERIMENTAL_CODE) $(GL_INCLUDES)
-vwebp_LDADD  = ../imageio/libexample_util.la ../imageio/libimageio_util.la
+vwebp_LDADD  = libexample_util.la ../imageio/libimageio_util.la
 vwebp_LDADD += ../src/demux/libwebpdemux.la $(GL_LIBS)
 
 if BUILD_LIBWEBPDECODER
