+++ examples/dwebp.c
@@ -40,7 +40,7 @@
 #endif
 
 #include "webp/decode.h"
-#include "../imageio/example_util.h"
+#include "../examples/example_util.h"
 #include "../imageio/imageio_util.h"
 #include "../imageio/webpdec.h"
 #include "./stopwatch.h"
