+++ examples/example_util.c
