+++ examples/gif2webp.c
@@ -26,7 +26,7 @@
 #include <gif_lib.h>
 #include "webp/encode.h"
 #include "webp/mux.h"
-#include "../imageio/example_util.h"
+#include "../examples/example_util.h"
 #include "../imageio/imageio_util.h"
 #include "./gifdec.h"
