+++ imageio/Makefile.am
@@ -1,12 +1,10 @@
 AM_CPPFLAGS += -I$(top_builddir)/src -I$(top_srcdir)/src
-noinst_LTLIBRARIES = libexample_util.la libimageio_util.la libimagedec.la
+noinst_LTLIBRARIES = libimageio_util.la libimagedec.la
 
 noinst_HEADERS =
 noinst_HEADERS += ../src/webp/decode.h
 noinst_HEADERS += ../src/webp/types.h
 noinst_HEADERS += ../examples/stopwatch.h
-
-libexample_util_la_SOURCES = example_util.c example_util.h
 
 libimageio_util_la_SOURCES = imageio_util.c imageio_util.h
 
@@ -19,5 +17,3 @@ libimagedec_la_SOURCES += webpdec.c webpdec.h
 libimagedec_la_SOURCES += wicdec.c wicdec.h
 libimagedec_la_CPPFLAGS = $(JPEG_INCLUDES) $(PNG_INCLUDES) $(TIFF_INCLUDES)
 libimagedec_la_CPPFLAGS += $(AM_CPPFLAGS) $(USE_EXPERIMENTAL_CODE)
-
-libexample_util_la_LIBADD = ../src/libwebp.la
