+++ makefile.unix
@@ -223,7 +223,7 @@ EX_FORMAT_DEC_OBJS = \
     imageio/webpdec.o \
 
 EX_UTIL_OBJS = \
-    imageio/example_util.o \
+    examples/example_util.o \
 
 GIFDEC_OBJS = \
     examples/gifdec.o \
@@ -310,7 +310,7 @@ HDRS = \
     src/webp/format_constants.h \
     $(HDRS_INSTALLED) \
 
-OUT_LIBS = imageio/libexample_util.a imageio/libimageio_util.a
+OUT_LIBS = examples/libexample_util.a imageio/libimageio_util.a
 OUT_LIBS += imageio/libimagedec.a src/libwebpdecoder.a
 OUT_LIBS += src/libwebp.a
 EXTRA_LIB = extras/libwebpextras.a
@@ -344,9 +344,9 @@ src/utils/bit_writer.o: src/utils/endian_inl.h
 	$(CC) $(CFLAGS) $(CPPFLAGS) -c $< -o $@
 
 examples/libanim_util.a: $(ANIM_UTIL_OBJS)
+examples/libexample_util.a: $(EX_UTIL_OBJS)
 examples/libgifdec.a: $(GIFDEC_OBJS)
 extras/libwebpextras.a: $(LIBWEBPEXTRA_OBJS)
-imageio/libexample_util.a: $(EX_UTIL_OBJS)
 imageio/libimagedec.a: $(EX_FORMAT_DEC_OBJS)
 imageio/libimageio_util.a: $(IMAGE_UTIL_OBJS)
 src/libwebpdecoder.a: $(LIBWEBPDECODER_OBJS)
@@ -365,25 +365,25 @@ examples/vwebp: examples/vwebp.o
 examples/webpmux: examples/webpmux.o
 
 examples/anim_diff: examples/libanim_util.a examples/libgifdec.a
-examples/anim_diff: src/demux/libwebpdemux.a imageio/libexample_util.a
+examples/anim_diff: src/demux/libwebpdemux.a examples/libexample_util.a
 examples/anim_diff: imageio/libimageio_util.a src/libwebp.a
 examples/anim_diff: EXTRA_LIBS += $(GIF_LIBS)
 examples/anim_diff: EXTRA_FLAGS += -DWEBP_HAVE_GIF
-examples/cwebp: imageio/libexample_util.a imageio/libimageio_util.a
+examples/cwebp: examples/libexample_util.a imageio/libimageio_util.a
 examples/cwebp: imageio/libimagedec.a src/libwebp.a
 examples/cwebp: EXTRA_LIBS += $(CWEBP_LIBS)
-examples/dwebp: imageio/libexample_util.a imageio/libimageio_util.a
+examples/dwebp: examples/libexample_util.a imageio/libimageio_util.a
 examples/dwebp: imageio/libimagedec.a src/libwebp.a
 examples/dwebp: EXTRA_LIBS += $(DWEBP_LIBS)
-examples/gif2webp: imageio/libexample_util.a imageio/libimageio_util.a
+examples/gif2webp: examples/libexample_util.a imageio/libimageio_util.a
 examples/gif2webp: examples/libgifdec.a src/mux/libwebpmux.a src/libwebp.a
 examples/gif2webp: EXTRA_LIBS += $(GIF_LIBS)
 examples/gif2webp: EXTRA_FLAGS += -DWEBP_HAVE_GIF
-examples/vwebp: imageio/libexample_util.a src/demux/libwebpdemux.a
+examples/vwebp: examples/libexample_util.a src/demux/libwebpdemux.a
 examples/vwebp: imageio/libimageio_util.a src/libwebp.a
 examples/vwebp: EXTRA_LIBS += $(GL_LIBS)
 examples/vwebp: EXTRA_FLAGS += -DWEBP_HAVE_GL
-examples/webpmux: imageio/libexample_util.a imageio/libimageio_util.a
+examples/webpmux: examples/libexample_util.a imageio/libimageio_util.a
 examples/webpmux: src/mux/libwebpmux.a src/libwebpdecoder.a
 
 extras/get_disto: extras/get_disto.o
